// apps/client/src/constants/ROUTES.ts
/* eslint-disable no-unused-vars */
export enum ROUTES {
  LOGIN = '/',
  REGISTER = '/register',
  DASHBOARD = '/dashboard',
  PROFILE_PAGE = '/dashboard/profile',
  ADD_PRODUCT_PAGE = '/dashboard/add',
  CHAGE_PRODUCT = '/dashboard/change/:id',
  PRODUCT_PAGE = '/dashboard/product/:id',
  NOT_FOUND_PAGE = '*',
}